backend/requirements.txt
flask
python-dotenv
SQLAlchemy>=2.0
psycopg[binary]
stripe
razorpay
paypalrestsdk